adapters: make the screen-on call to u-s-c synchronous
Normally, repowerd calls the unity-screen-compositor to unblank the
screen via DBus asynchronously before incrementally increase the
brightness. The problem is that, on some device (such as Fairphone 2),
the brightness is ignored if the screen is blanked. So, there is a race
condition between the brightness is incremented to the max and the
screen being actually unblanked.

This commit makes this call synchronous, making sure that the screen will
be unblanked completely before the brightness is set. A small timeout of
1 second is added to prevent unity-screen-compositor from blocking
repowerd for too long.