Use default config from other directory (#7)
This is a stupid one: repowerd first looks in powerd folder for default configs, and takes it from there. If then later no device-specific config is found in repowerd folder, it assumes that the default config there is of no use...

Most porters would go to repowerd folder (hence its name) for bindmounting the default config only to find out its not working.