Revert "core: Suspend inhibitions should affect automatic suspend only when display turns off due to inactivity"
On the phone, pressing power button should respect the suspend blocker.
For example, the phone should not sleep if it's playing music.

The same might not hold true for the desktop. However, repowerd already
prevents turning off display/suspend from power button on desktop anyway
[1]. Besides, repowerd requires more fixes to be able to run on desktop
properly. So, it's not our priority at the moment.

This reverts commit 8fd9f6b6a12f479cb4cd4422c541a8b2b8e31155.

[1] Sees src/adapters/default_state_machine_options.cpp, line 42.