debian: uses a link instead of a file copy for powerd compat config
For compatibility reason, repowerd also shipped a copy of the default
config file in the powerd's folder to be bind-mounted, in addition to
the one in repowerd's. However, due to how repowerd parses its configs,
only one of the config file will be parsed. If device's config is bind-
mounted to the wrong file, it'll be ignored, to a surprise.

This commit replace the compatibility copy with the compatibility link.
This means bind-mounting at one path will take affect at the other path
too. This fixes the compatibility issue without a surprise to the future
porters.