core: do disallow automatic suspend on client request
The phone might be woken up due to other source, and the client may want
the phone to stay awake for a little longer. So, don't assume that we
already disallow it and always disallow automatic suspend when a client
request not to be suspended as it's safe to call the function
repeatedly.

Fixes https://github.com/ubports/ubuntu-touch/issues/1556