Treat upower's OnBattery changing as a power source change
Previously, we only assumed a power change had happened when the upower
display device changed. However, it is possible for the device's power
source to change from battery to line without the display device
changing. Since we always want the display to turn on when plugging in
or unplugging a charger, even if the device's battery states didn't
change due to it, handle upower's OnBattery changing too.

Most of the logic needed for this was already around:
default_state_machine's handle_power_source_change called
is_using_battery_power in upower, which checks the OnBattery property.
All we had to do was set the pieces in motion.